requirements.txt
imaplib2==3.6
requests==2.31.0
pandas==2.0.3
openai==0.27.8
google-generativeai==0.1.0
python-dotenv==1.0.0